docker/requirements.txt
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
connexion==3.0.6
Flask==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
opensearch-py==2.4.2
packaging==23.2
pluggy==1.4.0
pytest==8.0.2
python-dateutil==2.8.2
python-multipart==0.0.9
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
schedule==1.2.1
setuptools==69.1.1
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
Werkzeug==3.0.3
